extract load script
